Moving Tracker.h to include
